feat: rename model_service to huggingface_service
